backend/requirements.txt
fastapi
uvicorn
semgrep
gitpython
python-dotenv
boto3
aiofiles
botocore